refactor: do fetch call in dashboard component
